:art: move table row into table file
